refactor(api): extract messages API to separate module
Move message-related handlers (create_message, list_messages) from
sessions.rs to a new messages.rs module for better code organization.
